+++ vllm/v1/utils.py
@@ -16,12 +16,12 @@ import torch
 from torch.autograd.profiler import record_function
 
 import vllm.envs as envs
-from vllm.lite_profiler import combine_contexts, lite_profiler
 from vllm.logger import init_logger
 from vllm.usage.usage_lib import (UsageContext, is_usage_stats_enabled,
                                   usage_message)
 from vllm.utils import (get_open_port, get_open_zmq_ipc_path, get_tcp_uri,
                         kill_process_tree)
+from vllm.utils.lite_profiler import combine_contexts, lite_profiler
 
 if TYPE_CHECKING:
     import numpy as np
